1.订单支付服务 createPayment 创建支付订单服务 2.三方支付回调 callbackPayment 保存订单支付状态(保存成功则发送消息给MQ,让其他服务订阅并作出相应的业务操作) 3.优化日志消费Mapper的字段类型等问题

+++ shop-pojo/src/main/java/com/my/shop/pojo/Payment.java
@@ -18,7 +18,7 @@ public class Payment implements Serializable {
     //'支付金额'
     private Integer pay_amount;
     //'是否已支付 0否 1是'
-    private BigInteger is_paid;
+    private Integer is_paid;
 
     public BigInteger getId() {
         return id;
@@ -44,11 +44,11 @@ public class Payment implements Serializable {
         this.pay_amount = pay_amount;
     }
 
-    public BigInteger getIs_paid() {
+    public Integer getIs_paid() {
         return is_paid;
     }
 
-    public void setIs_paid(BigInteger is_paid) {
+    public void setIs_paid(Integer is_paid) {
         this.is_paid = is_paid;
     }
 }
